rename PluginHandler -> ViewHandler
