+++ bundles/statistics/statsgrid/handler/ViewHandler.js
